pyproject.toml
[project]
name = "2048-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "google-generativeai>=0.8.5",
    "numpy>=2.2.5",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "wandb>=0.19.10",
]
